+++ desktop_version/CMakeLists.txt
@@ -117,7 +117,10 @@ set(VVV_SRC
 	../third_party/physfs/extras/physfsrwops.c
 )
 if(NOT CUSTOM_LEVEL_SUPPORT STREQUAL "DISABLED")
-	list(APPEND VVV_SRC src/editor.cpp)
+	list(APPEND VVV_SRC src/CustomLevels.cpp)
+	if(NOT CUSTOM_LEVEL_SUPPORT STREQUAL "NO_EDITOR")
+		LIST(APPEND VVV_SRC src/Editor.cpp)
+	endif()
 endif()
 if(STEAM)
 	list(APPEND VVV_SRC src/SteamNetwork.c)
